backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
prisma==0.12.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
google-generativeai==0.4.0
python-multipart==0.0.9
